requirements.txt
Flask==2.0.2
mysql-connector-python==8.0.27
gunicorn==20.1.0


